AnswerId is now a uint for MessagePollAnswer.cs ( And its JSON variant )

+++ NetCord/MessagePollAnswer.cs
@@ -6,6 +6,6 @@ namespace NetCord;
 public class MessagePollAnswer(JsonMessagePollAnswer jsonModel, ulong guildId, RestClient client) : IJsonModel<JsonMessagePollAnswer>
 {
     JsonMessagePollAnswer IJsonModel<JsonMessagePollAnswer>.JsonModel => jsonModel;
-    public ulong AnswerId => jsonModel.AnswerId;
+    public uint AnswerId => jsonModel.AnswerId;
     public MessagePollMedia PollMedia { get; } = new(jsonModel.PollMedia, guildId, client);
 }
